Remove all sorting stuff until real sorting, extract seasonsTableCell and TwoColumnCell components
