+++ src/components/tables/SeasonsTable.tsx
@@ -5,124 +5,27 @@ import { SeasonsTableData } from "@/state/useSeasonsData";
 import { Tooltip, TooltipProvider, TooltipTrigger } from "@/components/ui/Tooltip";
 import { TooltipContent, TooltipPortal } from "@radix-ui/react-tooltip";
 import { seasonColumns, SortColumn } from "@/pages/explorer/SeasonsExplorer";
-import { TokenValue } from "@/classes/TokenValue";
 import React, { useEffect, useMemo, useRef, useState } from "react";
 import { areEqual, ListChildComponentProps, VariableSizeList } from "react-window";
 import useIsMobile from "@/hooks/display/useIsMobile";
 import { calculateCropScales, convertDeltaDemandToPercentage } from "@/utils/convert";
 import { caseIdToDescriptiveText } from "@/utils/utils";
-
-enum SeasonsTableCellType {
-  Default = "default",
-  TwoColumn = "twoColumn",
-}
-
-interface SeasonsTableCellProps {
-  cellType?: SeasonsTableCellType;
-  columnKey: string;
-  value?: any;
-  subValue?: any;
-  className?: string;
-  hoverContent?: any;
-}
-
-const TwoColumnCell = ({
-  className,
-  value,
-  subValue,
-  hoverContent,
-}: { className: string; value: any; subValue: any; hoverContent?: any }): JSX.Element => {
-  const cellContent = (
-    <div className="flex flex-col items-end">
-      <span>{value}</span>
-      <span className="text-xs break-keep text-pinto-gray-4">{subValue}</span>
-    </div>
-  );
-  return hoverContent ? (
-    <TableCell className={className}>
-      <TooltipProvider>
-        <Tooltip delayDuration={0}>
-          <TooltipTrigger>{cellContent}</TooltipTrigger>
-          <TooltipPortal>
-            <TooltipContent className="z-50 overflow-hidden" side="right" sideOffset={4}>
-              {hoverContent}
-            </TooltipContent>
-          </TooltipPortal>
-        </Tooltip>
-      </TooltipProvider>
-    </TableCell>
-  ) : (
-    <TableCell className={className}>{cellContent}</TableCell>
-  );
-};
-
-const nonHideableFields = ["season"];
+import { SeasonsTableCell, SeasonsTableCellType } from "./SeasonsTableCell";
 
 interface SeasonsTableProps {
   seasonsData: SeasonsTableData[];
   hiddenFields: string[];
   hideColumn: (id: string) => void;
-  toggleColumn: (id: string) => void;
-  sortedColumn: SortColumn;
-  setSortedColumn: (column: SortColumn) => void;
 }
 
-// const getRotationClassNameForColumn = (column: SortColumn, id: string) => {
-//   if (column.column === id) {
-//     return column.dir === 'asc' ? 'rotate-90' : '-rotate-90';
-//   }
-//   return '-rotate-90';
-// }
+export const nonHideableFields = ['season']
 
-// const getRotationForColumn = (column: SortColumn, id: string) => {
-//   // desc is default, otherwise flip the direction
-//   if (column.column === id) {
-//     return column.dir === 'asc' ? 'desc' : 'asc';
-//   }
-//   return 'desc';
-// }
-
-export const SeasonsTable = ({ seasonsData, hiddenFields, hideColumn, sortedColumn, setSortedColumn }: SeasonsTableProps) => {
+export const SeasonsTable = ({ seasonsData, hiddenFields, hideColumn }: SeasonsTableProps) => {
 
   const tableRef = useRef<HTMLTableElement>(null);
   const isMobile = useIsMobile();
   const [height, setHeight] = useState(500);
   const paginationPadding = isMobile ? 62 : 64
-
-
-  const SeasonsTableCell = ({
-    cellType = SeasonsTableCellType.Default,
-    columnKey,
-    value,
-    subValue,
-    className = "text-right",
-    hoverContent,
-  }: SeasonsTableCellProps) => {
-    if (hiddenFields.includes(columnKey)) return null;
-    const column = seasonColumns.find(c => c.id === columnKey);
-    const additionalClasses = column?.classes;
-    switch (cellType) {
-      case SeasonsTableCellType.TwoColumn:
-        return <TwoColumnCell className={`${className} ${additionalClasses}`} value={value} subValue={subValue} hoverContent={hoverContent} />;
-      default:
-        return hoverContent ? (
-          <TableCell className={`${className} ${additionalClasses}`}>
-            <TooltipProvider>
-              <Tooltip delayDuration={0}>
-                <TooltipTrigger>{value}</TooltipTrigger>
-                <TooltipPortal>
-                  <TooltipContent className="z-50 overflow-hidden" side="right" sideOffset={4}>
-                    {hoverContent}
-                  </TooltipContent>
-                </TooltipPortal>
-              </Tooltip>
-            </TooltipProvider>
-          </TableCell>
-        ) : (
-          <TableCell className={`${className} ${additionalClasses}`}>{value}</TableCell>
-        );
-    }
-  };
 
   const calculatedWidth = useMemo(() => {
     return seasonColumns.reduce((acc, column) => {
@@ -160,9 +63,9 @@ export const SeasonsTable = ({ seasonsData, hiddenFields, hideColumn, sortedColu
     if (data.season <= 3) {
       return (
         <TableRow key={data.season} style={style} noHoverMute>
-          {seasonColumns.map(({ id, name, classes }) => {
+          {seasonColumns.map(({ id }) => {
             return (
-              <SeasonsTableCell key={id} columnKey={id} value={id === 'season' ? data.season : 'N/A'} />
+              <SeasonsTableCell key={id} columnKey={id} value={id === 'season' ? data.season : 'N/A'} hiddenFields={hiddenFields} />
             )
           })}
         </TableRow>
@@ -170,60 +73,71 @@ export const SeasonsTable = ({ seasonsData, hiddenFields, hideColumn, sortedColu
     }
     return (
       <TableRow key={data.season} style={style} noHoverMute>
-        <SeasonsTableCell className="text-left" columnKey="season" value={data.season} />
+        <SeasonsTableCell className="text-left" columnKey="season" value={data.season} hiddenFields={hiddenFields} />
         <SeasonsTableCell
           columnKey="instantDeltaP"
           value={`${data.instDeltaB.toNumber() > 0 ? "+" : ""}${data.instDeltaB.toHuman("short")}`}
+          hiddenFields={hiddenFields}
         />
         <SeasonsTableCell
           columnKey="twaDeltaP"
           value={`${data.twaDeltaB.toNumber() > 0 ? "+" : ""}${data.twaDeltaB.toHuman("short")}`}
+          hiddenFields={hiddenFields}
         />
         <SeasonsTableCell
           columnKey="pintoSupply"
           value={`${data.deltaBeans.toNumber() > 0 ? "+" : ""}${data.deltaBeans.toHuman("short")}`}
+          hiddenFields={hiddenFields}
         />
         <SeasonsTableCell
           columnKey="totalSoil"
           value={data.issuedSoil.toHuman("short")}
+          hiddenFields={hiddenFields}
         />
         <SeasonsTableCell
           cellType={SeasonsTableCellType.TwoColumn}
           columnKey="soilSown"
           value={data.deltaSownBeans.toHuman("short")}
           subValue={`Total: ${data.sownBeans.toHuman("short")}`}
+          hiddenFields={hiddenFields}
         />
         <SeasonsTableCell
           columnKey="timeSown"
           value={data.blocksToSoldOutSoil}
+          hiddenFields={hiddenFields}
         />
         <SeasonsTableCell
           columnKey="price"
           value={`$${data.price.toNumber().toFixed(4)}`}
+          hiddenFields={hiddenFields}
         />
         <SeasonsTableCell
           cellType={SeasonsTableCellType.TwoColumn}
           columnKey="twaPrice"
           value={`$${data.twaPrice.toHuman("short")}`}
           subValue={caseIdToDescriptiveText(data.caseId, "price")}
+          hiddenFields={hiddenFields}
         />
         <SeasonsTableCell
           cellType={SeasonsTableCellType.TwoColumn}
           columnKey="l2sr"
           value={`${data.l2sr.toHuman("short")}%`}
           subValue={caseIdToDescriptiveText(data.caseId, "l2sr")}
+          hiddenFields={hiddenFields}
         />
         <SeasonsTableCell
           cellType={SeasonsTableCellType.TwoColumn}
           columnKey="podRate"
           value={`${data.podRate.toHuman("short")}%`}
           subValue={caseIdToDescriptiveText(data.caseId, "pod_rate")}
+          hiddenFields={hiddenFields}
         />
         <SeasonsTableCell
           cellType={SeasonsTableCellType.TwoColumn}
           columnKey="deltaDemand"
           value={convertDeltaDemandToPercentage(data.deltaPodDemand.toNumber())}
           subValue={caseIdToDescriptiveText(data.caseId, 'soil_demand')}
+          hiddenFields={hiddenFields}
         // hoverContent={
         //   <DeltaDemandChart />
         // }
@@ -238,10 +152,13 @@ export const SeasonsTable = ({ seasonsData, hiddenFields, hideColumn, sortedColu
                 {deltaCropScalar})
               </span>
             </>
-          } />
+          }
+          hiddenFields={hiddenFields}
+        />
         <SeasonsTableCell
           columnKey="cropRatio"
           value={`${cropRatio}%`}
+          hiddenFields={hiddenFields}
         />
         <SeasonsTableCell
           cellType={SeasonsTableCellType.TwoColumn}
@@ -256,6 +173,7 @@ export const SeasonsTable = ({ seasonsData, hiddenFields, hideColumn, sortedColu
             </>
           }
           subValue={"Temperature"}
+          hiddenFields={hiddenFields}
         />
       </TableRow>
     );
